+++ packages/renderer/src/dialogs/csv-paste-dialog.ts
@@ -3,8 +3,7 @@ import { customElement, state } from 'lit/decorators.js';
 import { getContext } from '../app-context.js';
 import { parseCsv } from '../plugins/csv-import.js';
 import { slugTable } from '../util/ids.js';
-import { ctrlEnterSubmits, dialogChromeStyles } from './dialog-chrome.js';
-import { makeDialogDraggable } from './draggable.js';
+import { ctrlEnterSubmits, dialogChromeStyles, makeDialogDraggable } from '@cawoodm/lit-dialogs';
 import { watchDialogDirty } from '../chrome/dirty-guard.js';
 
 /**